3. appliquer les labels/requirements.txt
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.100.0
